Refactor user model and update related imports; remove unused user.py file
