+++ app/api/v1/movies.py
@@ -1,4 +1,3 @@
-from typing import List
 from fastapi import Depends
 from sqlalchemy.orm import Session
 from app.core.database import get_db
