+++ app/core/config.py
@@ -6,7 +6,7 @@ class Settings(BaseSettings):
     ACCESS_TOKEN_EXPIRE_MINUTES: int  = 60
     REFRESH_TOKEN_EXPIRE_DAYS: int  = 7
     ALGORITHM: str  = ""
-
+    
     class Config:
         env_file = ".env"
